feat(observations): initial implementation
motivation:
---
i wanted a library-level way to track certain 'entities' that are generated upon the 'put' to each store. specifically this comes out of llm analysis, i want to track for example whenever some says the word "apple", every version of the input needs to keep a ref to this & update

approach:
---
opus-4.5 vibe coded this.

testing:
---
haven't actually validated the tests that it did add.